backend/requirements.txt
Flask==2.3.2
Flask-SocketIO==5.3.6
eventlet==0.33.3
Flask-Cors==3.0.10
Flask-Bcrypt==1.0.1
Flask-RESTx==1.3.0
pymongo==4.7.2
python-dotenv==1.0.1
gunicorn==21.2.0
requests==2.31.0
openai==1.30.1
python-binance==1.0.19
dnspython==2.6.1
PyJWT==2.8.0
setuptools